Moved update profile function to utils
